+++ packages/furo-data-input/test/furo-data-money-input.test.js
@@ -117,4 +117,30 @@ describe('furo-data-money-input', () => {
       dataInput.shadowRoot.getElementById('select').dispatchEvent(customEvent);
     }, 0);
   });
+
+  it('should set meta via response meta', done => {
+    console.log('those tests are base on the mockdata/experiment/1/get.json');
+
+    host._FBPAddWireHook('--hts', () => {
+      entityObject.addEventListener(
+        'data-injected',
+        () => {
+          setTimeout(() => {
+            console.log(dataInput.shadowRoot.getElementById('input'));
+            assert.equal(
+              dataInput.shadowRoot.getElementById('input').getAttribute('disabled'),
+              null,
+            );
+            assert.equal(
+              dataInput.shadowRoot.getElementById('select').getAttribute('disabled'),
+              null,
+            );
+            done();
+          }, 55);
+        },
+        { once: true },
+      );
+    });
+    deeplink.qpIn({ exp: 1 });
+  });
 });