+++ packages/furo-data/test/furo-entity-agent.test.js
@@ -27,15 +27,7 @@ describe('furo-entity-agent', () => {
     await dataObject.updateComplete;
   });
 
-  it('should be a furo-entity-agent', done => {
-    // keep this test on top, so you can recognize a wrong asignment
-    assert.equal(entityAgent.nodeName.toLowerCase(), 'furo-entity-agent');
-    assert.equal(dataObject.nodeName.toLowerCase(), 'furo-data-object');
-    done();
-  });
-
   it('request payload should only have writeable or required properties inside', done => {
-    entityAgent.loadOnHtsIn = true;
     entityAgent.setAttribute('service', 'ExperimentService');
     dataObject.setAttribute('type', 'experiment.ExperimentEntity');
 
@@ -57,17 +49,9 @@ describe('furo-entity-agent', () => {
     dataObject.addEventListener('data-injected', () => {
       dataObject.data.data.description._value = 'description';
       dataObject.data.data.id._value = '1';
-
-      entityAgent.save();
+      entityAgent.bindRequestData(dataObject.data.data);
+      entityAgent.create();
     });
-
-    entityAgent.addEventListener(
-      'response',
-      r => {
-        dataObject.injectRaw(r.detail);
-      },
-      { once: true },
-    );
 
     entityAgent.htsIn([
       {
@@ -78,13 +62,233 @@ describe('furo-entity-agent', () => {
         service: 'ExperimentService',
       },
       {
-        href: 'http://httpbin.org/anything',
+        href: 'https://httpbin.org/anything',
         method: 'POST',
         rel: 'create',
         type: 'experiment.ExperimentEntity',
         service: 'ExperimentService',
       },
     ]);
+    dataObject.injectRaw({
+      data: {
+        description: 'experiment data for testing',
+        display_name: 'display_name',
+        id: '1',
+        furo_data_checkbox_input: true,
+        furo_data_text_input: 'hallo test',
+        furo_data_number_input: 12.55,
+        furo_data_time_input: '17:34',
+        furo_data_color_input: '#e318ed',
+        furo_data_textarea_input: 'hallo , this is textarea input',
+        furo_data_date_input: '2019-02-22',
+        furo_data_date_input_google: {
+          day: 31,
+          display_name: '31.12.2020',
+          month: 12,
+          year: 2020,
+        },
+        furo_data_range_input: 31,
+        furo_data_bool_icon: false,
+        type_property: [
+          {
+            code: 'c0a7f550-0fbe-4046-8fa9-60c86327b6b1',
+            data: {
+              '@type': 'type.googleapis.com/furo.StringProperty',
+              data: '01012020',
+            },
+            display_name: 'Vertragsbeginn',
+            id: 'a396a165-ebf9-4bd5-870f-9f09a929f9f9',
+            meta: {
+              fields: {
+                'data.data': {
+                  constraints: {
+                    required: {
+                      is: 'true',
+                      message: 'Bitte ausfüllen!',
+                    },
+                  },
+                  meta: {
+                    label: 'Vertragsbeginn',
+                    readonly: true,
+                  },
+                },
+                meta: {
+                  meta: {
+                    readonly: true,
+                  },
+                },
+              },
+            },
+          },
+          {
+            data: {
+              '@type': 'xx/furo.NumberProperty',
+              data: 34.23,
+              display_name: '34.23',
+            },
+            display_name: 'Display',
+            id: 'bb',
+            code: 'option',
+            meta: {
+              fields: {
+                'data.data': {
+                  constraints: {
+                    required: {
+                      is: 'true',
+                      message: 'Bitte ausfüllen!',
+                    },
+                  },
+                  meta: {
+                    label: 'Float',
+                    hint: 'Type in int',
+                    readonly: true,
+                  },
+                },
+              },
+            },
+          },
+          {
+            data: {
+              '@type': 'xx/furo.IntegerProperty',
+              data: 342,
+              display_name: '342',
+            },
+            display_name: 'Display',
+            id: 'op33t',
+            code: 'option',
+            meta: {
+              fields: {
+                'data.data': {
+                  meta: {
+                    label: 'Integer',
+                    hint: 'Type in int',
+                  },
+                },
+              },
+            },
+          },
+          {
+            data: {
+              '@type': 'xx/furo.StringOptionProperty',
+              id: 'bb',
+              display_name: 'Display',
+            },
+            display_name: 'Display',
+            id: '444',
+            code: 'option',
+            meta: {
+              fields: {
+                data: {
+                  meta: {
+                    label: 'Please select',
+                    hint: 'datehint is data',
+                    repeated: false,
+                    options: {
+                      list: [
+                        {
+                          id: '',
+                          display_name: 'Please select',
+                        },
+                        {
+                          id: 'aa',
+                          display_name: 'The display a',
+                        },
+                        {
+                          id: 'bb',
+                          display_name: 'The display b',
+                        },
+                      ],
+                    },
+                  },
+                },
+              },
+            },
+          },
+          {
+            data: {
+              '@type': 'xx/furo.StringProperty',
+              data: '32 32 23',
+              display_name: 'a date',
+            },
+            id: 'bv3',
+            code: 'date',
+            meta: {
+              fields: {
+                'data.data': {
+                  meta: {
+                    label: 'repeated fields',
+                    hint: 'this is data',
+                  },
+                },
+              },
+            },
+          },
+        ],
+        single_type_property: {
+          data: {
+            '@type': 'xx/google.type.Date',
+            day: 8,
+            month: 11,
+            year: 2022,
+            display_name: '8.11.2022',
+          },
+          display_name: 'a date',
+          id: 'date',
+          code: 'date',
+          meta: {
+            fields: {
+              data: {
+                meta: {
+                  label: 'Additional fields',
+                  hint: 'this is data',
+                },
+                constraints: {
+                  min: {
+                    is: '2019-09-09',
+                    message: 'to small',
+                  },
+                },
+              },
+            },
+          },
+        },
+        repstring: ['AAA', 'BBBB', 'CCCC'],
+        furo_data_money_input: {
+          currency_code: 'CHF',
+          units: 3333,
+          nanos: 75100000,
+        },
+      },
+      meta: {
+        fields: {
+          'data.furo_data_color_input': {
+            meta: {
+              readonly: true,
+            },
+          },
+          'data.type_property': {
+            meta: {
+              label: 'Label from response',
+            },
+          },
+        },
+      },
+      links: [
+        {
+          href: 'https://httpbin.org/anything',
+          method: 'POST',
+          rel: 'create',
+          type: 'experiment.ExperimentEntity',
+          service: 'ExperimentService',
+        },
+      ],
+    });
+  });
+  it('should be a furo-entity-agent', done => {
+    // keep this test on top, so you can recognize a wrong asignment
+    assert.equal(entityAgent.nodeName.toLowerCase(), 'furo-entity-agent');
+    assert.equal(dataObject.nodeName.toLowerCase(), 'furo-data-object');
+    done();
   });
 
   it('Required fields must always be fully transmitted.', done => {
@@ -124,6 +328,13 @@ describe('furo-entity-agent', () => {
         href: '/mockdata/experiments/1/get.json',
         method: 'GET',
         rel: 'self',
+        type: 'experiment.ExperimentEntity',
+        service: 'ExperimentService',
+      },
+      {
+        href: 'http://httpbin.org/anything',
+        method: 'POST',
+        rel: 'create',
         type: 'experiment.ExperimentEntity',
         service: 'ExperimentService',
       },
@@ -184,7 +395,7 @@ describe('furo-entity-agent', () => {
         service: 'ExperimentService',
       },
       {
-        href: 'http://httpbin.org/anything',
+        href: 'https://httpbin.org/anything',
         method: 'PATCH',
         rel: 'update',
         type: 'experiment.ExperimentEntity',
@@ -214,7 +425,7 @@ describe('furo-entity-agent', () => {
         service: 'TaskService',
       },
       {
-        href: 'http://httpbin.org/anything',
+        href: 'https://httpbin.org/anything',
         method: 'PUT',
         rel: 'update',
         type: 'task.TaskEntity',
@@ -297,7 +508,7 @@ describe('furo-entity-agent', () => {
     });
     entityAgent.htsIn([
       {
-        href: 'http://httpbin.org/anything',
+        href: 'https://httpbin.org/anything',
         method: 'DELETE',
         rel: 'delete',
         type: 'task.TaskEntity',
@@ -332,7 +543,7 @@ describe('furo-entity-agent', () => {
     });
     entityAgent.htsIn([
       {
-        href: 'http://httpbin.org/anything',
+        href: 'https://httpbin.org/anything',
         method: 'PUT',
         rel: 'update',
         type: 'task.TaskEntity',
@@ -351,7 +562,7 @@ describe('furo-entity-agent', () => {
     });
     entityAgent.htsIn([
       {
-        href: 'http://httpbin.org/anything',
+        href: 'https://httpbin.org/anything',
         method: 'POST',
         rel: 'create',
         type: 'task.TaskEntity',
@@ -373,7 +584,7 @@ describe('furo-entity-agent', () => {
     });
     entityAgent.htsIn([
       {
-        href: 'http://httpbin.org/anything',
+        href: 'https://httpbin.org/anything',
         method: 'POST',
         rel: 'create',
         type: 'task.TaskEntity',
@@ -403,7 +614,7 @@ describe('furo-entity-agent', () => {
         service: 'TaskService',
       },
       {
-        href: 'http://httpbin.org/anything',
+        href: 'https://httpbin.org/anything',
         method: 'PUT',
         rel: 'update',
         type: 'task.TaskEntity',
@@ -435,7 +646,7 @@ describe('furo-entity-agent', () => {
         service: 'TaskService',
       },
       {
-        href: 'http://httpbin.org/anything',
+        href: 'https://httpbin.org/anything',
         method: 'PATCH',
         rel: 'update',
         type: 'task.TaskEntity',
@@ -496,5 +707,88 @@ describe('furo-entity-agent', () => {
     entityAgent.setAttribute('service', 'TaskService');
     assert.equal(entityAgent.load(), false);
     done();
+  });
+
+  it('Accept header should be set ', done => {
+    entityAgent.setAttribute('service', 'TaskService');
+
+    entityAgent.addEventListener('hts-updated', () => {
+      const request = entityAgent._makeRequest({
+        href: '/mockdata/tasks/1/get.json',
+        method: 'GET',
+        rel: 'self',
+        type: 'task.TaskEntity',
+        service: 'TaskService',
+      });
+      assert.equal(
+        request.headers.get('Accept'),
+        'application/task.TaskEntity+json, application/json;q=0.9',
+      );
+      done();
+    });
+
+    entityAgent.htsIn([
+      {
+        href: '/mockdata/tasks/1/get.json',
+        method: 'GET',
+        rel: 'self',
+        type: 'task.TaskEntity',
+        service: 'TaskService',
+      },
+    ]);
+  });
+
+  it('should accept clearing of QueryParams ', done => {
+    entityAgent.setAttribute('service', 'TaskService');
+    entityAgent.updateQp({ compact: true });
+    entityAgent.updateQp({ test: '5' });
+    entityAgent.clearQp();
+    /**
+     * Register hook on wire --triggerLoad to
+     *
+     */
+    entityAgent._FBPAddWireHook('--triggerLoad', req => {
+      assert.equal(req.url.indexOf('compact=true') > 0, false);
+      assert.equal(req.url.indexOf('test=5') > 0, false);
+      done();
+    });
+
+    entityAgent.htsIn([
+      {
+        href: '/mockdata/tasks/1/get.json',
+        method: 'GET',
+        rel: 'self',
+        type: 'task.TaskEntity',
+        service: 'TaskService',
+      },
+    ]);
+    entityAgent.load();
+  });
+
+  it('should rebuild request url with via updateQp setted qp and the previously existed qp together in url ', done => {
+    entityAgent.setAttribute('service', 'TaskService');
+    entityAgent.updateQp({ compact: true });
+    entityAgent.updateQp({ test: '5' });
+    /**
+     * Register hook on wire --triggerLoad to
+     *
+     */
+    entityAgent._FBPAddWireHook('--triggerLoad', req => {
+      assert.equal(req.url.indexOf('compact=true') > 0, true);
+      assert.equal(req.url.indexOf('test=5') > 0, true);
+      assert.equal(req.url.indexOf('previousqp=xyz') > 0, true);
+      done();
+    });
+
+    entityAgent.htsIn([
+      {
+        href: '/mockdata/tasks/1/get.json?previousqp=xyz',
+        method: 'GET',
+        rel: 'self',
+        type: 'task.TaskEntity',
+        service: 'TaskService',
+      },
+    ]);
+    entityAgent.load();
   });
 });