+++ packages/furo-framework/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.12](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/framework@1.3.11...@furo/framework@1.3.12) (2020-06-29)
+
+**Note:** Version bump only for package @furo/framework
+
+
+
+
+
+## [1.3.11](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/framework@1.3.10...@furo/framework@1.3.11) (2020-06-22)
+
+**Note:** Version bump only for package @furo/framework
+
+
+
+
+
+## [1.3.10](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/framework@1.3.9...@furo/framework@1.3.10) (2020-06-16)
+
+**Note:** Version bump only for package @furo/framework
+
+
+
+
+
+## [1.3.9](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/framework@1.3.8...@furo/framework@1.3.9) (2020-06-08)
+
+**Note:** Version bump only for package @furo/framework
+
+
+
+
+
+## [1.3.8](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/framework@1.3.7...@furo/framework@1.3.8) (2020-06-04)
+
+**Note:** Version bump only for package @furo/framework
+
+
+
+
+
 ## [1.3.7](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/framework@1.3.6...@furo/framework@1.3.7) (2020-05-01)
 
 **Note:** Version bump only for package @furo/framework